Fixed issue where the Track response format
For the API method `user.gettoptracks`, the format of the Track response contains a field named `streamable` which is JSON object type, while the same in the Track response for the method `user.getrecenttracks` is a numeric boolean.

Therefore, created a separate Track class for gettoptracks until I figure out a way to simplify the implementation.